requirements.txt
cachetools==5.5.0
colorama==0.4.6
DateTime==5.5
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
numpy==2.2.0
oauthlib==3.2.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytz==2024.2
requests-oauthlib==2.0.0
rsa==4.9
tzdata==2024.2
zope.interface==7.2
